requirements.txt
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
numpy==1.22.1
pandas==1.4.0
parsel==1.6.0
priority==2.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
queuelib==1.6.2
requests==2.27.1
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
zope.interface==5.4.0
